Simplify how StorageServiceDAO's are created across the different storage backends (#224)
- externalized thread pool and refresh interval configuration

+++ front50-core/src/main/groovy/com/netflix/spinnaker/front50/model/application/DefaultApplicationDAO.java
@@ -29,7 +29,7 @@ import java.util.Map;
 public class DefaultApplicationDAO extends StorageServiceSupport<Application> implements ApplicationDAO {
   public DefaultApplicationDAO(StorageService service,
                                Scheduler scheduler,
-                               int refreshIntervalMs,
+                               long refreshIntervalMs,
                                Registry registry) {
     super(ObjectType.APPLICATION, service, scheduler, refreshIntervalMs, registry);
   }
